refactor with Configuration namespace
