+++ PH.WorkingDaysAndTime/PH.WorkingDaysAndTimeUtility.UnitTest/WorkDaySpanUnitTest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using PH.WorkingDaysAndTimeUtility.Configuration;
 
 namespace PH.WorkingDaysAndTimeUtility.UnitTest
 {
