+++ PH.WorkingDaysAndTime/PH.WorkingDaysAndTimeUtility.UnitTest/WorkingDaysAndTimeUtilityUnitTest.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using PH.WorkingDaysAndTimeUtility.Configuration;
 
 namespace PH.WorkingDaysAndTimeUtility.UnitTest
 {
