+++ PH.WorkingDaysAndTime/PH.WorkingDaysAndTimeUtility/Configuration/EasterSunday.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace PH.WorkingDaysAndTimeUtility
+namespace PH.WorkingDaysAndTimeUtility.Configuration
 {
     /// <summary>
     /// Easter Sunday HoliDay
