+++ PH.WorkingDaysAndTime/PH.WorkingDaysAndTimeUtility/WorkingDaysAndTimeUtility.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
+using PH.WorkingDaysAndTimeUtility.Configuration;
 
 namespace PH.WorkingDaysAndTimeUtility
 {
